utilities: Disable debug tool on Windows.
The current implementation of the debug tool is not compatible with VC compiler.
This patch disables the debug tool on Windows builds as it is not needed.
Of course it would be better if the code could be fixed to work with VC++,
but this is a quick fix to get the Windows build working.

Signed-off-by: Frank Wagner <frank.wagner@dbosoft.eu>